Try to implement keymap by creating my own
that maps straight from ASCII.

I'm getting a new and different error from this, but that seems like
there might be a way to handle it? If this "is-a" shared-memory pool,
or if we have a shared-memory pool which is sized at least this big?

That is: the compositor isn't going to allocate memory for us, we have
to allocate it (take the quota) and then share it with the compositor.
That would make sense. A lot of sense, actually, like a _lot_ of sense.